fully tested

+++ RateCustomers.py
@@ -2,6 +2,7 @@ import sqlite3 as db
 import csv
 import sys
 
+
 def create_tables():
 
     conn = db.connect("Orders.db")
@@ -13,7 +14,7 @@ def create_tables():
 
     cur.execute("create table barcodes (barcode, order_id)")
 
-    with open('barcodes.csv', 'r') as fin:  # `with` statement available in 2.5+
+    with open('barcodes.csv', 'r') as fin:
         # csv.DictReader uses first line in file for column headings by default
         dr = csv.DictReader(fin)  # comma is default delimiter
         to_db = [(i['barcode'], i['order_id']) for i in dr]
@@ -47,24 +48,25 @@ def executeNonQuery(query):
     conn.close()
 
 def validate_barcodes():
-    '''validate no dupicate bar codes'''
+    # validate no duplicate bar codes
 
     dic, conn = executeQuery("Select count(barcode), barcode from barcodes group by barcode having count(barcode) > 1")
 
-    print(sys.stderr, "Deleting duplicate barcodes found for: "+", ".join(str(v) for k, v in dic))
+    if dic is not None:
+        print("Deleting duplicate barcodes found for: " + ", ".join(str(v) for k, v in dic), file=sys.stderr)
 
-    executeNonQuery("delete from barcodes where barcode in"
-                "(select barcode from(Select count(barcode),barcode from barcodes "
-                "group by barcode having count(barcode) > 1))")
+        executeNonQuery("delete from barcodes where barcode in"
+                        "(select barcode from(Select count(barcode),barcode from barcodes "
+                        "group by barcode having count(barcode) > 1))")
 
 def validate_orders():
-    '''validate no order without bar code'''
+    # validate no order without bar code
 
     res, conn = executeQuery("select order_id from orders o where not exists "
                 "(select 1 from barcodes b where b.order_id = o.order_id )")
     conn.close()
     if res is not None:
-        print(sys.stderr, "Deleting orders with no barcode: " + ", ".join(str(i[0]) for i in res))
+        print("Deleting orders with no barcode: " + ", ".join(str(i[0]) for i in res), file=sys.stderr)
 
         executeNonQuery("delete from orders where order_id in("
                     "select order_id from orders o where not exists "
@@ -84,17 +86,18 @@ def write_to_csv(data):
         writer.writerows(data)
         f.close()
 
-
-def print_top_five():
-    ''''''
+def get_top_five():
     res, conn = executeQuery("select o.customer_id, count(b.barcode) from orders o, barcodes b "
                  "where b.order_id = o.order_id "
                  "group by o.customer_id "
                  "order by count(b.barcode) desc")
-    print("Top customers: "+ ", ".join(str(res[i][0]) for i in range(0, 5)))
     conn.close()
+    return (str(res[i][0]) for i in range(0, 5))
 
-
+def get_unused_barcodes():
+    res, conn = executeQuery("select count(barcode) from barcodes where order_id = ''")
+    conn.close()
+    return res[0][0]
 
 
 def main():
@@ -103,15 +106,10 @@ def main():
     validate_orders()
     dataset = get_customer_to_order()
     write_to_csv(dataset)
-    print_top_five()
+    print("The top 5 customers are: " + ", ".join(get_top_five()))
+    print("There are: " + str(get_unused_barcodes()) + " unused barcodes.")
 
 
-
-
-
-    '''join order to barcodes on order_id'''
-    '''generate to output file'''
-    '''group by customer amount of tickets, print top 5'''
 
 if __name__ == "__main__":
     main()
